Pushed All Changes since 28/02/13
Since this date, we have added a second grapple hook function, added
cables to the grappling hooks (when used) so Gerald is not floating,
added panels to the game so Gerald can only use them to grapple hook
with the second function (So you cant cheat). We have also added
floating logs which sink when you stand on them. The level has also been
added to since. Some other changes but I cant remember :)